changed scale factor from int to float. we can now run at arbitrary resolutions. simpler than expected :D

+++ src/MuiConfig.h
@@ -72,7 +72,7 @@ namespace mui{
 		static bool useRetinaAssets;
 		
 		// pixels per pixel.
-		static int scaleFactor;
+		static float scaleFactor;
 		
 		// path to the data folder (containing the mui folder)
 		static Poco::Path dataPath;
